feature: 게시글 상세 기능 구현

+++ CatchMate/data/src/main/java/com/catchmate/data/dto/BoardWriteResponseDTO.kt
@@ -1,5 +1,5 @@
 package com.catchmate.data.dto
 
 data class BoardWriteResponseDTO(
-    val boardId: Int,
+    val boardId: Long,
 )
