+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/BoardWriteResponse.kt
@@ -1,5 +1,5 @@
 package com.catchmate.domain.model
 
 data class BoardWriteResponse(
-    val boardId: Int,
+    val boardId: Long,
 )
